# decp-platform/docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: DECP Platform API
  description: |
    **Department Engagement & Career Platform** — REST API documentation.

    ## Authentication
    All protected endpoints require a Bearer JWT token in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```

    Access tokens expire after **15 minutes**. Use the `/auth/refresh` endpoint to obtain a new pair.

    ## Response Format
    All responses follow the canonical envelope:
    ```json
    {
      "success": true,
      "message": "Human-readable description",
      "data": {},
      "meta": { "page": 1, "limit": 10, "total": 50 }
    }
    ```

    ## Base URL
    - Development: `http://localhost:3000/api/v1`
    - Production: `https://api.decp.edu/api/v1`
  version: 1.0.0
  contact:
    name: DECP Platform Team
    email: admin@decp.edu
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development (via API Gateway)
  - url: https://api.decp.edu/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 15
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
        nextCursor:
          type: string
          nullable: true
          description: ISO timestamp for cursor-based pagination

    # ── User ──────────────────────────────────────────────────────────────────
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        content:
          type: string
          maxLength: 5000
        type:
          type: string
          enum: [text, image, video, document]
        mediaUrls:
          type: array
          items:
            type: string
        likes:
          type: integer
        comments:
          type: integer
        shares:
          type: integer
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Comment ───────────────────────────────────────────────────────────────
...
paths:
...
  /posts:
    get:
      tags: [Feed]
      summary: Get public feed (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
        - name: cursor
          in: query
          description: ISO timestamp for cursor-based pagination
          schema:
            type: string
        - name: userId
          in: query
          description: Filter posts by author
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Feed retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
    post:
      tags: [Feed]
      summary: Create a new post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  maxLength: 5000
                type:
                  type: string
                  enum: [text, image, video, document]
                isPublic:
                  type: boolean
                  default: true
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      post:
                        $ref: '#/components/schemas/Post'
        '401':
          description: Not authenticated
...